Move old `PropertiesConfiguration` to external module
The old properties configuration prevents the new one from having the
same configuration factory order.

This commit moves it to a new `log4j-config-properties-legacy` module.
