+++ log4j-config-properties-legacy/src/main/java/org/apache/logging/log4j/core/config/properties/PropertiesConfiguration.java
