+++ log4j-config-properties-legacy/src/test/java/org/apache/logging/log4j/core/config/properties/PropertiesConfigurationTest.java
