+++ log4j-config-properties-legacy/src/test/java/org/apache/logging/log4j/core/config/properties/PropertiesRoutingAppenderTest.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.logging.log4j.core.appender.routing;
+package org.apache.logging.log4j.core.config.properties;
 
 import static org.junit.jupiter.api.Assertions.*;
