+++ log4j-config-properties-legacy/src/test/resources/LoggerLevelAppenderTest.properties
