+++ log4j-config-properties-legacy/src/test/resources/LoggerLevelSysPropsAppenderTest.properties
