+++ log4j-config-properties-legacy/src/test/resources/RootLoggerLevelAppenderTest.properties
