+++ log4j-config-properties-legacy/src/test/resources/log4j-routing.properties
