+++ log4j-config-properties-legacy/src/test/resources/log4j-test1.properties
