+++ log4j-config-properties-legacy/src/test/resources/log4j2-properties-trailing-space-on-level.properties
