--- log4j-core-test/src/test/resources/RollingRandomAccessFileAppenderLocationPropsTest.properties
@@ -1,37 +1,0 @@
-#
-# Licensed to the Apache Software Foundation (ASF) under one or more
-# contributor license agreements.  See the NOTICE file distributed with
-# this work for additional information regarding copyright ownership.
-# The ASF licenses this file to you under the Apache License, Version 2.0
-# (the "License"); you may not use this file except in compliance with
-# the License.  You may obtain a copy of the License at
-#
-#      http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-#
-
-status = off
-name = PropertiesConfigTest
-
-appenders = rolling
-
-appender.rolling.type = RollingRandomAccessFile
-appender.rolling.name = RollingRandomAccessFile
-appender.rolling.fileName = target/RollingRandomAccessFileAppenderLocationPropsTest.log
-appender.rolling.filePattern = target/RollingProps/test-%d{MM-dd-yyyy}-%i.log.gz
-appender.rolling.append = false
-appender.rolling.immediateFlush = true;
-appender.rolling.layout.type = PatternLayout
-appender.rolling.layout.pattern = %d %p %c{1.} [%t] %X{aKey} %m %location %ex%n
-appender.rolling.policies.type = Policies
-appender.rolling.policies.time.type = TimeBasedTriggeringPolicy
-
-rootLogger.level = info
-rootLogger.includeLocation = false
-rootLogger.appenderRefs = rolling
-rootLogger.appenderRef.rolling.ref = RollingRandomAccessFile
