--- log4j-core-test/src/test/resources/log4j-test2.properties
@@ -1,59 +1,0 @@
-#
-# Licensed to the Apache Software Foundation (ASF) under one or more
-# contributor license agreements.  See the NOTICE file distributed with
-# this work for additional information regarding copyright ownership.
-# The ASF licenses this file to you under the Apache License, Version 2.0
-# (the "License"); you may not use this file except in compliance with
-# the License.  You may obtain a copy of the License at
-#
-#      http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-#
-
-status = off
-name = PropertiesConfigTest
-monitorInterval = 1
-
-property.filename = target/test-properties.log
-
-filter.threshold.type = ThresholdFilter
-filter.threshold.level = debug
-
-appender.console.name = STDOUT
-appender.console.type = Console
-appender.console.layout.type = PatternLayout
-appender.console.layout.pattern = %m%n
-
-appender.file.name = File
-appender.file.type = File
-appender.file.fileName = ${filename}
-appender.file.bufferedIO = false
-appender.file.layout.type = PatternLayout
-appender.file.layout.pattern = %d %p %C{1.} [%t] %m%n
-
-appender.list.name = List
-appender.list.type = List
-appender.list.filter.threshold.type = ThresholdFilter
-appender.list.filter.threshold.level = error
-
-logger.test1.name = org.apache.logging.log4j.test1
-logger.test1.level = debug
-logger.test1.additivity = false
-logger.test1.filter.mdc.type = ThreadContextMapFilter
-logger.test1.filter.mdc.pair.type = KeyValuePair
-logger.test1.filter.mdc.pair.key = test
-logger.test1.filter.mdc.pair.value = 123
-logger.test1.appenderRef.console.ref = STDOUT
-
-logger.test2.name = org.apache.logging.log4j.test2
-logger.test2.level = debug
-logger.test2.additivity = false
-logger.test2.appenderRef.file.ref = File
-
-rootLogger.level = error
-rootLogger.appenderRef.console.ref = STDOUT
