+++ pom.xml
@@ -238,6 +238,7 @@
     <module>log4j-appserver</module>
     <module>log4j-config-jackson</module>
     <module>log4j-config-properties</module>
+    <module>log4j-config-properties-legacy</module>
     <module>log4j-config-yaml</module>
     <module>log4j-core</module>
     <module>log4j-core-its</module>
@@ -386,6 +387,12 @@
       <dependency>
         <groupId>org.apache.logging.log4j</groupId>
         <artifactId>log4j-config-properties</artifactId>
+        <version>${project.version}</version>
+      </dependency>
+
+      <dependency>
+        <groupId>org.apache.logging.log4j</groupId>
+        <artifactId>log4j-config-properties-legacy</artifactId>
         <version>${project.version}</version>
       </dependency>
