+++ src/changelog/.3.x.x/create_config_properties.xml
@@ -1,9 +1,9 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <entry xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
        xmlns="http://logging.apache.org/log4j/changelog"
-       xsi:schemaLocation="http://logging.apache.org/log4j/changelog https://logging.apache.org/log4j/changelog-0.1.2.xsd"
+       xsi:schemaLocation="http://logging.apache.org/log4j/changelog https://logging.apache.org/log4j/changelog-0.1.3.xsd"
        type="added">
   <description format="asciidoc">
-    Add a new configuration factory based on `jackson-dataformat-properties`.
+    Add a new properties configuration factory based on `jackson-dataformat-properties`.
   </description>
 </entry>
