+++ src/changelog/.3.x.x/create_config_properties_legacy.xml
@@ -1,0 +1,9 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<entry xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+       xmlns="http://logging.apache.org/log4j/changelog"
+       xsi:schemaLocation="http://logging.apache.org/log4j/changelog https://logging.apache.org/log4j/changelog-0.1.2.xsd"
+       type="changed">
+  <description format="asciidoc">
+    Moves legacy `2.x` properties configuration factory to `log4j-config-properties-legacy`.
+  </description>
+</entry>
